# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Movie Booking Application.",
        "title": "Bioskop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/promos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get list of active promos (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promos"
                ],
                "summary": "Get all promos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/movie-app_internal_domain.Promo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a new promo code (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promos"
                ],
                "summary": "Create new promo",
                "parameters": [
                    {
                        "description": "Promo Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movie-app_internal_delivery_http_dto_request.CreatePromoRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/movie-app_internal_domain.Promo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/promos/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update promo details (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promos"
                ],
                "summary": "Update promo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/movie-app_internal_delivery_http_dto_request.UpdatePromoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/movie-app_internal_domain.Promo"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete/Remove a promo (Admin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Promos"
                ],
                "summary": "Delete promo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Promo UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/reports/revenue": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "See daily or monthly revenue (Admin Only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Get revenue report",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report mode: 'day' or 'month'",
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/movie-app_pkg_utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/movie-app_internal_delivery_http_dto_response.DailyRevenueResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/reports/revenue/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download revenue report as CSV file (Admin Only)",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "Reports"
                ],
                "summary": "Export revenue CSV",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Report mode: 'day' or 'month'",
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "movie-app_internal_delivery_http_dto_request.CreatePromoRequest": {
            "type": "object",
            "required": [
                "code",
                "discount_type",
                "discount_value",
                "valid_until"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "discount_type": {
                    "type": "string",
                    "enum": [
                        "percentage",
                        "fixed"
                    ]
                },
                "discount_value": {
                    "type": "number"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
        "movie-app_internal_delivery_http_dto_request.UpdatePromoRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "discount_type": {
                    "type": "string",
                    "enum": [
                        "percentage",
                        "fixed"
                    ]
                },
                "discount_value": {
                    "type": "number"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
        "movie-app_internal_delivery_http_dto_response.DailyRevenueResponse": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "total_amount": {
                    "type": "number"
                },
                "transaction_count": {
                    "type": "integer"
                }
            }
        },
...
        "movie-app_internal_domain.Promo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "discount_type": {
                    "description": "'percentage' or 'fixed'",
                    "type": "string"
                },
                "discount_value": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
        "movie-app_pkg_utils.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}